Refactor runner tests into fixture-driven suites
